updating lab4. Changed data type of max to size_t line 96.

+++ lab4/main.cpp
@@ -93,9 +93,9 @@ pair<string, string> parseInput(string input) {
 
 //uses insertAtBegin to push each element of value1 and value2 onto thier respective LinkedList.
 void addNodes(pair <string, string> values, LinkedList &num1, LinkedList &num2) {
-  unsigned long size = max(values.first.size(), values.second.size());
+  size_t size = max(values.first.size(), values.second.size());
   
-  for (unsigned long i = 0; i < size; i++ ) {
+  for (size_t i = 0; i < size; i++ ) {
     if (i >= values.first.size()) {
       num1.insertAtEnd(0);
     }
